Move into a block, redirect the /extract path to the tab
