+++ apps/merge/src/state/extract.ts
@@ -1,0 +1,13 @@
+import type { GeoBbox2D } from "@osmix/shared/types"
+import { atom } from "jotai"
+import { atomWithStorage } from "jotai/utils"
+import { DEFAULT_EXTRACT_BBOX } from "../lib/extract-bbox"
+
+export const activeTabAtom = atomWithStorage<string>(
+	"@osmix:merge:activeTab",
+	"Inspect",
+)
+
+export const extractBboxAtom = atom<GeoBbox2D>(DEFAULT_EXTRACT_BBOX)
+
+export const extractSyncMapToBboxAtom = atom(false)
